feat: add support for token helper
Add support for token helper, a command line tool to obtain a token.

A token helper is an executable, set in the user's `.npmrc` which
outputs an auth token. This can be used in situations where the
`authToken` is not a constant value, but is something that refreshes
regularly, where a script or other tool can use an existing refresh
token to obtain a new access token.

The configuration for the path to the helper must be an absolute path,
with no arguments. In order to be secure, it is _only_ permitted to set
this value in the user `.npmrc`, otherwise a project could place a value
in a project local `.npmrc` and run arbitrary executables. For this
reason, the userConfig is added as a third parameter to the
function, to enable only allowing this value from that `.npmrc`.

A similar feature is available in many similar tools. The implementation
here is modelled after the `vault` (vaultproject.io) implementation
- https://github.com/hashicorp/vault/blob/main/command/token/helper_external.go